# Optimize inserting into memtable with multiple thread.