--- a/db/db_impl/db_impl.h
+++ b/db/db_impl/db_impl.h
@@ -159,8 +159,8 @@ class DBImpl : public DB {
   virtual Status Write(const WriteOptions& options,
                        WriteBatch* updates) override;
 
-  using DB::MultiThreadWrite;
-  virtual Status MultiThreadWrite(const WriteOptions& options, const std::vector<WriteBatch*> &updates) override;
+  using DB::MultiBatchWrite;
+  virtual Status MultiBatchWrite(const WriteOptions& options, const std::vector<WriteBatch*> &updates) override;
 
   using DB::Get;
   virtual Status Get(const ReadOptions& options,
@@ -1058,10 +1058,10 @@ class DBImpl : public DB {
                    size_t batch_cnt = 0,
                    PreReleaseCallback* pre_release_callback = nullptr);
 
-  Status MultiThreadWriteImpl(const WriteOptions& write_options,
-                         const autovector<WriteBatch*>& my_batch, WriteCallback* callback,
-                         uint64_t* log_used = nullptr, uint64_t log_ref = 0,
-                         uint64_t* seq_used = nullptr);
+  Status MultiBatchWriteImpl(const WriteOptions &write_options,
+                             const autovector<WriteBatch *> &my_batch, WriteCallback *callback,
+                             uint64_t *log_used = nullptr, uint64_t log_ref = 0,
+                             uint64_t *seq_used = nullptr);
 
   Status PipelinedWriteImpl(const WriteOptions& options, WriteBatch* updates,
                             WriteCallback* callback = nullptr,