--- a/db/db_impl/db_impl_open.cc
+++ b/db/db_impl/db_impl_open.cc
@@ -128,6 +128,11 @@ DBOptions SanitizeOptions(const std::string& dbname, const DBOptions& src) {
     result.enable_multi_thread_write = false;
   }
 
+  if (result.enable_multi_thread_write) {
+    result.enable_pipelined_write = true;
+    result.allow_concurrent_memtable_write = true;
+  }
+
 #ifndef ROCKSDB_LITE
   ImmutableDBOptions immutable_db_options(result);
   if (!IsWalDirSameAsDBPath(&immutable_db_options)) {