--- a/db/db_impl/db_impl_open.cc
+++ b/db/db_impl/db_impl_open.cc
@@ -123,14 +123,13 @@ DBOptions SanitizeOptions(const std::string& dbname, const DBOptions& src) {
     result.avoid_flush_during_recovery = false;
   }
 
-  // multi thread write do not support two-write-que or write in pipeline
-  if (result.two_write_queues) {
+  // multi thread write do not support two-write-que or without allow_concurrent_memtable_write
+  if (result.two_write_queues || !result.allow_concurrent_memtable_write) {
     result.enable_multi_thread_write = false;
   }
 
   if (result.enable_multi_thread_write) {
     result.enable_pipelined_write = true;
-    result.allow_concurrent_memtable_write = true;
   }
 
 #ifndef ROCKSDB_LITE