--- a/db/db_impl/db_impl_write.cc
+++ b/db/db_impl/db_impl_write.cc
@@ -60,6 +60,134 @@ Status DBImpl::WriteWithCallback(const WriteOptions& write_options,
 }
 #endif  // ROCKSDB_LITE
 
+
+Status DBImpl::MultiThreadWrite(const WriteOptions& options, const std::vector<WriteBatch*> &updates) {
+  if (immutable_db_options_.enable_multi_thread_write && !updates.empty()) {
+    autovector<WriteBatch*> batches;
+    for (auto w: updates) {
+      batches.push_back(w);
+    }
+    return MultiThreadWriteImpl(options, batches, nullptr, nullptr);
+  } else {
+    Status status;
+    for (auto b: updates) {
+      status = WriteImpl(options, b, nullptr, nullptr);
+      if (!status.ok()) {
+        return status;
+      }
+    }
+    return status;
+  }
+}
+
+Status DBImpl::MultiThreadWriteImpl(const WriteOptions& write_options,
+                         const autovector<WriteBatch*>& updates, WriteCallback* callback,
+                         uint64_t* log_used, uint64_t log_ref,
+                         uint64_t* seq_used) {
+  StopWatch write_sw(env_, immutable_db_options_.statistics.get(), DB_WRITE);
+  WriteThread::Writer writer(write_options, updates, callback, log_ref);
+  write_thread_.JoinBatchGroup(&writer);
+
+  WriteContext write_context;
+  PERF_TIMER_GUARD(write_pre_and_post_process_time);
+  bool ignore_missing_faimly = write_options.ignore_missing_column_families;
+  if (writer.state == WriteThread::STATE_GROUP_LEADER) {
+    if (writer.callback && !writer.callback->AllowWriteBatching()) {
+      write_thread_.WaitForMemTableWriters();
+    }
+    WriteThread::WriteGroup wal_write_group;
+    mutex_.Lock();
+    bool need_log_sync = !write_options.disableWAL && write_options.sync;
+    bool need_log_dir_sync = need_log_sync && !log_dir_synced_;
+    PERF_TIMER_STOP(write_pre_and_post_process_time);
+    writer.status = PreprocessWrite(write_options, &need_log_sync, &write_context);
+    PERF_TIMER_START(write_pre_and_post_process_time);
+    log::Writer *log_writer = logs_.back().writer;
+    mutex_.Unlock();
+
+    // This can set non-OK status if callback fail.
+    last_batch_group_size_ =
+            write_thread_.EnterAsBatchGroupLeader(&writer, &wal_write_group);
+    const SequenceNumber current_sequence = write_thread_.UpdateLastSequence(versions_->LastSequence()) + 1;
+    size_t total_count = 0;
+    size_t total_byte_size = 0;
+    size_t valid_batches = 0;
+    auto stats = default_cf_internal_stats_;
+    if (writer.status.ok()) {
+      SequenceNumber next_sequence = current_sequence;
+      for (auto w : wal_write_group) {
+        if (w->CheckCallback(this)) {
+          if (w->ShouldWriteToMemtable()) {
+            w->sequence = next_sequence;
+            size_t count = WriteBatchInternal::Count(w->batches);
+            next_sequence += count;
+            total_count += count;
+          }
+          valid_batches += 1;
+          total_byte_size = WriteBatchInternal::AppendedByteSize(
+                  total_byte_size, WriteBatchInternal::ByteSize(w->batches));
+        }
+      }
+      if (writer.disable_wal) {
+        has_unpersisted_data_.store(true, std::memory_order_relaxed);
+      }
+      write_thread_.UpdateLastSequence(current_sequence + total_count - 1);
+      if (!write_options.disableWAL) {
+        PERF_TIMER_GUARD(write_wal_time);
+        stats->AddDBStats(InternalStats::kIntStatsWriteDoneBySelf, 1);
+        RecordTick(stats_, WRITE_DONE_BY_SELF, 1);
+        if (wal_write_group.size > 1) {
+          stats->AddDBStats(InternalStats::kIntStatsWriteDoneByOther,
+                  wal_write_group.size - 1);
+          RecordTick(stats_, WRITE_DONE_BY_OTHER, wal_write_group.size - 1);
+        }
+        writer.status = WriteToWAL(wal_write_group, log_writer, log_used,
+                                           need_log_sync, need_log_dir_sync, current_sequence);
+      }
+    }
+    if (!writer.CallbackFailed()) {
+      WriteStatusCheck(writer.status);
+    }
+
+    if (need_log_sync) {
+      mutex_.Lock();
+      MarkLogsSynced(logfile_number_, need_log_dir_sync, writer.status);
+      mutex_.Unlock();
+    }
+    write_thread_.ExitAsBatchGroupLeader(wal_write_group, writer.status);
+  }
+  if (writer.state == WriteThread::STATE_MEMTABLE_WRITER_LEADER) {
+    PERF_TIMER_GUARD(write_memtable_time);
+    assert(writer.ShouldWriteToMemtable());
+    WriteThread::WriteGroup memtable_write_group;
+    write_thread_.EnterAsMemTableWriter(&writer, &memtable_write_group);
+    assert(immutable_db_options_.allow_concurrent_memtable_write);
+    auto version_set = versions_->GetColumnFamilySet();
+    memtable_write_group.running.store(0);
+    for (auto it = memtable_write_group.begin(); it != memtable_write_group.end(); ++it) {
+      if (!it.writer->ShouldWriteToMemtable()) {
+        continue;
+      }
+      WriteBatchInternal::AsyncInsertInto(it.writer, it.writer->sequence, version_set,
+                                          &flush_scheduler_, &trim_history_scheduler_, ignore_missing_faimly, this,
+                                          write_options.memtable_insert_hint_per_batch, write_pool_.get());
+    }
+    while (memtable_write_group.running.load(std::memory_order_acquire) > 0) {
+      bool succ = write_pool_->StealOneJobAndRun();
+      if (!succ) {
+        std::this_thread::yield();
+      }
+    }
+    versions_->SetLastSequence(memtable_write_group.last_sequence);
+    write_thread_.ExitAsMemTableWriter(&writer, memtable_write_group);
+  }
+  if (seq_used != nullptr) {
+    *seq_used = writer.sequence;
+  }
+
+  return writer.status;
+}
+
 // The main write queue. This is the only write queue that updates LastSequence.
 // When using one write queue, the same sequence also indicates the last
 // published sequence.
@@ -96,6 +224,19 @@ Status DBImpl::WriteImpl(const WriteOptions& write_options,
     return Status::NotSupported(
         "pipelined_writes is not compatible with unordered_write");
   }
+
+  if (immutable_db_options_.unordered_write &&
+      immutable_db_options_.enable_multi_thread_write) {
+    return Status::NotSupported(
+        "multi_thread_writes is not compatible with unordered_write");
+  }
+
+  if (immutable_db_options_.enable_multi_thread_write) {
+    autovector<WriteBatch*> updates;
+    updates.push_back(my_batch);
+    return MultiThreadWriteImpl(write_options, updates, callback, log_used, log_ref, seq_used);
+  }
+
   // Otherwise IsLatestPersistentState optimization does not make sense
   assert(!WriteBatchInternal::IsLatestPersistentState(my_batch) ||
          disable_memtable);
@@ -927,12 +1068,12 @@ WriteBatch* DBImpl::MergeBatch(const WriteThread::WriteGroup& write_group,
   *write_with_wal = 0;
   auto* leader = write_group.leader;
   assert(!leader->disable_wal);  // Same holds for all in the batch group
-  if (write_group.size == 1 && !leader->CallbackFailed() &&
-      leader->batch->GetWalTerminationPoint().is_cleared()) {
+  if (write_group.size == 1 && !leader->CallbackFailed() && leader->batches.size() == 1 &&
+      leader->batches[0]->GetWalTerminationPoint().is_cleared()) {
     // we simply write the first WriteBatch to WAL if the group only
     // contains one batch, that batch should be written to the WAL,
     // and the batch is not wanting to be truncated
-    merged_batch = leader->batch;
+    merged_batch = leader->batches[0];
     if (WriteBatchInternal::IsLatestPersistentState(merged_batch)) {
       *to_be_cached_state = merged_batch;
     }
@@ -944,11 +1085,12 @@ WriteBatch* DBImpl::MergeBatch(const WriteThread::WriteGroup& write_group,
     merged_batch = tmp_batch;
     for (auto writer : write_group) {
       if (!writer->CallbackFailed()) {
-        WriteBatchInternal::Append(merged_batch, writer->batch,
-                                   /*WAL_only*/ true);
-        if (WriteBatchInternal::IsLatestPersistentState(writer->batch)) {
-          // We only need to cache the last of such write batch
-          *to_be_cached_state = writer->batch;
+        for (auto b : writer->batches) {
+          WriteBatchInternal::Append(merged_batch, b, /*WAL_only*/ true);
+          if (WriteBatchInternal::IsLatestPersistentState(b)) {
+            // We only need to cache the last of such write batch
+            *to_be_cached_state = b;
+          }
         }
         (*write_with_wal)++;
       }
@@ -1594,7 +1736,7 @@ Status DBImpl::SwitchMemtable(ColumnFamilyData* cfd, WriteContext* context) {
 
   // In case of pipelined write is enabled, wait for all pending memtable
   // writers.
-  if (immutable_db_options_.enable_pipelined_write) {
+  if (immutable_db_options_.enable_pipelined_write || immutable_db_options_.enable_multi_thread_write) {
     // Memtable writers may call DB::Get in case max_successive_merges > 0,
     // which may lock mutex. Unlocking mutex here to avoid deadlock.
     mutex_.Unlock();