--- a/db/write_batch_internal.h
+++ b/db/write_batch_internal.h
@@ -23,6 +23,7 @@ namespace rocksdb {
 class MemTable;
 class FlushScheduler;
 class ColumnFamilyData;
+class ColumnFamilySet;
 
 class ColumnFamilyMemTables {
  public:
@@ -117,6 +118,8 @@ class WriteBatchInternal {
   // Return the number of entries in the batch.
   static uint32_t Count(const WriteBatch* batch);
 
+  static uint32_t Count(const autovector<WriteBatch*> b);
+
   // Set the count for the number of entries in the batch.
   static void SetCount(WriteBatch* batch, uint32_t n);
 
@@ -137,6 +140,14 @@ class WriteBatchInternal {
 
   static size_t ByteSize(const WriteBatch* batch) {
     return batch->rep_.size();
+  }
+
+  static size_t ByteSize(const autovector<WriteBatch*> batches) {
+    size_t sz = 0;
+    for (auto b: batches) {
+      sz += b->rep_.size();
+    }
+    return sz;
   }
 
   static Status SetContents(WriteBatch* batch, const Slice& contents);
@@ -190,6 +201,11 @@ class WriteBatchInternal {
                            bool seq_per_batch = false, size_t batch_cnt = 0,
                            bool batch_per_txn = true,
                            bool hint_per_batch = false);
+
+  static void AsyncInsertInto(WriteThread::Writer* writer, SequenceNumber sequence,
+        ColumnFamilySet* version_set, FlushScheduler* flush_scheduler,
+        TrimHistoryScheduler* trim_history_scheduler,
+        bool ignore_missing_column_families, DB* db, bool hint_per_batch, ThreadPool* pool);
 
   static Status Append(WriteBatch* dst, const WriteBatch* src,
                        const bool WAL_only = false);