--- a/db/write_thread.h
+++ b/db/write_thread.h
@@ -22,7 +22,9 @@
 #include "rocksdb/status.h"
 #include "rocksdb/types.h"
 #include "rocksdb/write_batch.h"
+#include "rocksdb/threadpool.h"
 #include "util/autovector.h"
+
 
 namespace rocksdb {
 
@@ -114,6 +116,7 @@ class WriteThread {
   // Information kept for every waiting writer.
   struct Writer {
     WriteBatch* batch;
+    autovector<WriteBatch*> batches;
     bool sync;
     bool no_slowdown;
     bool disable_wal;
@@ -163,6 +166,29 @@ class WriteThread {
           disable_wal(write_options.disableWAL),
           disable_memtable(_disable_memtable),
           batch_cnt(_batch_cnt),
+          pre_release_callback(_pre_release_callback),
+          log_used(0),
+          log_ref(_log_ref),
+          callback(_callback),
+          made_waitable(false),
+          state(STATE_INIT),
+          write_group(nullptr),
+          sequence(kMaxSequenceNumber),
+          link_older(nullptr),
+          link_newer(nullptr) {
+      batches.push_back(batch);
+    }
+
+    Writer(const WriteOptions& write_options, const autovector<WriteBatch*>& _batch,
+           WriteCallback* _callback, uint64_t _log_ref,
+           PreReleaseCallback* _pre_release_callback = nullptr)
+        : batch(nullptr),
+          batches(_batch),
+          sync(write_options.sync),
+          no_slowdown(write_options.no_slowdown),
+          disable_wal(write_options.disableWAL),
+          disable_memtable(false),
+          batch_cnt(0),
           pre_release_callback(_pre_release_callback),
           log_used(0),
           log_ref(_log_ref),
@@ -275,6 +301,7 @@ class WriteThread {
   //
   // Writer* w:        Writer to be executed as part of a batch group
   void JoinBatchGroup(Writer* w);
+  void AwaitWriterState(Writer* w);
 
   // Constructs a write batch group led by leader, which should be a
   // Writer passed to JoinBatchGroup on the current thread.
@@ -349,6 +376,14 @@ class WriteThread {
   // Remove the dummy writer and wake up waiting writers
   void EndWriteStall();
 
+  void SetWritePool(const std::shared_ptr<ThreadPool>& pool) {
+    write_pool_ = pool;
+  }
+
+  void ResetWritePool() {
+    write_pool_.reset();
+  }
+
  private:
   // See AwaitState.
   const uint64_t max_yield_usec_;
@@ -359,6 +394,7 @@ class WriteThread {
 
   // Enable pipelined write to WAL and memtable.
   const bool enable_pipelined_write_;
+  const bool enable_multi_thread_write_;
 
   // The maximum limit of number of bytes that are written in a single batch
   // of WAL or memtable write. It is followed when the leader write size
@@ -387,6 +423,7 @@ class WriteThread {
   // on the writer queue
   port::Mutex stall_mu_;
   port::CondVar stall_cv_;
+  std::shared_ptr<ThreadPool> write_pool_;
 
   // Waits for w->state & goal_mask using w->StateMutex().  Returns
   // the state that satisfies goal_mask.