--- a/include/rocksdb/db.h
+++ b/include/rocksdb/db.h
@@ -384,7 +384,7 @@ class DB {
   // Note: consider setting options.sync = true.
   virtual Status Write(const WriteOptions& options, WriteBatch* updates) = 0;
 
-  virtual Status MultiThreadWrite(const WriteOptions& options, const std::vector<WriteBatch*> &updates) {
+  virtual Status MultiBatchWrite(const WriteOptions& options, const std::vector<WriteBatch*> &updates) {
     Status s;
     for (auto w: updates) {
       s = Write(options, w);