--- a/include/rocksdb/threadpool.h
+++ b/include/rocksdb/threadpool.h
@@ -47,6 +47,8 @@ class ThreadPool {
   virtual void SubmitJob(const std::function<void()>&) = 0;
   // This moves the function in for efficiency
   virtual void SubmitJob(std::function<void()>&&) = 0;
+
+  virtual bool StealOneJobAndRun() = 0;
 };
 
 // NewThreadPool() is a function that could be used to create a ThreadPool