## Design

* Although RocksDB could use `allow_concurrent_memtable_write` to allow write-thread to insert data into memtable together, which belong to one group. However, if one thread call `Write` while there is still one leader inserting into memtable , it must wait for new leader to pull it into another write-group.
 * `unordered_write` has almost solved this problem. But there is flaw, that is, no longer provide a consistent snapshot read.
 * So I have one new interface `MultiThreadWrite`,  which allow customer to write multiple `WriteBatch` at once.  Every thread could help leader to insert into memtable when they wait for next leader. Of course, it provides **snapshot read**.
* I also add a `ThreadPool` to help write thread for some special considerations. For example, application with RocksDB may have multiple write-thread, and they are not balanced. And this application must write data into RocksDB in order, because it is a system requiring linear consistency, such as `Raft` or `Paxos`.
* use `enable_multi_thread_write` to enable this feature.

## Bench

There is the script of benchmark.
```
mkdir -p data
rm -rf data/*

./db_bench --db=./data --disable_wal=true --enable_pipelined_write=true --use_multi_thread_write=true --memtable_write_pool_size=0 --key_size=20 --value_size=100 --write_buffer_size=4096000000  --benchmarks=fillrandom --batch_size=512 --num=1000000 --threads=6
rm -rf data/*
./db_bench --db=./data --disable_wal=true --enable_pipelined_write=true --key_size=20 --value_size=100 --write_buffer_size=4096000000  --benchmarks=fillrandom --batch_size=512 --num=1000000 --threads=6
rm -rf data/*
./db_bench --db=./data --disable_wal=true --enable_pipelined_write=false --unordered_write=true --key_size=20 --value_size=100 --write_buffer_size=4096000000  --benchmarks=fillrandom --batch_size=512 --num=1000000 --threads=6
```

option\thread | 6 |   12  |  24
-|-|-|-
multi_thread_write | 214 MB/s | 361 MB/s | 524 MB/s |
pipelined_write | 131 MB/s | 191 MB/s | 285 MB/s |
unordered_write | 220MB/s | 383 MB/s | 545 MB/s |

## Conclusion

* Obliviously, multi_thread_write could have similar performance compare to unordered_write, and both of them are much better than pipelined_write.