@maysamyabandeh 2019-09-25
This is a very interesting direction. Thanks for looking into this. I am still not sure what is the gist of the design and where the efficiency is coming from. But in the meanwhile I gave the patch a try with the same benchmark that was provided in unordered_write PR. The results that I get are quite different. Have you tried these benchmarks? Do you where the discrepancy is coming from?
```
conf/wal                    WAL          no-WAL
base                          275           445  MB/s
unordered_write              439           1135
pull                          70              71
```
```
TEST_TMPDIR=/dev/shm/ ./db_bench_pull --benchmarks=fillrandom --threads=32 --num=10000000 -max_write_buffer_number=16 --max_background_jobs=64 --batch_size=8 --writes=3000000 -level0_file_num_compaction_trigger=99999 --level0_slowdown_writes_trigger=99999 --level0_stop_writes_trigger=99999  -disable_auto_compactions  --unordered_write=0 --enable_pipelined_write=true --use_multi_thread_write=true --memtable_write_pool_size=0 --disable_wal=1
```