@yiwu-arbug 2019-09-25
@Little-Wallace may have more insight, but my understanding is the patch is useful in scenario when there's only few threads concurrent writing (e.g. 6 threads) but with large write batch. Due to the way write group is formed, for the default write impl, a write group may consist of fewer writer than the number of user write threads, so the concurrency is lower than expected; while with the patch, we can simply set a larger write thread pool size to achieve higher memtable write concurrency.

For our use case we have relatively large write batch and we need atomic write for the whole batch, so it's difficult to split the batch into several and concurrent write from multiple user threads.