@maysamyabandeh 2019-09-25
Thanks @yiwu-arbug So I understand that the patch is to optimize for large batch_size 512.
Have you tried running the benchmark with --transaction_db which provide immutable snapshot. When I do I get results similar to what the patch does.
```
24t+nowal+unordered_write          743 MB/s
24t+nowal+unordered_write+txn      676 MB/s
24t+nowal+patch                    681 MB/s
```
```
TEST_TMPDIR=/dev/shm/ ./db_bench_pull --disable_wal=true --enable_pipelined_write=false --unordered_write=true --key_size=20 --value_size=100 --write_buffer_size=4096000000  --benchmarks=fillrandom --batch_size=512 --num=1000000 --threads=24

TEST_TMPDIR=/dev/shm/ ./db_bench_pull --disable_wal=true --enable_pipelined_write=false --unordered_write=true --key_size=20 --value_size=100 --write_buffer_size=4096000000  --benchmarks=fillrandom --batch_size=512 --num=1000000 --threads=24 --transaction_db

TEST_TMPDIR=/dev/shm/ ./db_bench_pull --disable_wal=true --enable_pipelined_write=true --use_multi_thread_write=true --memtable_write_pool_size=0 --key_size=20 --value_size=100 --write_buffer_size=4096000000  --benchmarks=fillrandom --batch_size=512 --num=1000000 --threads=24
```