@Little-Wallace 2019-09-26
Thanks @maysamyabandeh for your attention.
Emmm， I means that this pr would perform better than normal pipelined-write. But it is still slightly slower than unordered-write. 
What I have optimized is just normal write, and it could provide consistant read. In our project , unordered-write would make our development more difficult, so we use pipelined-write until now.  I thought there may be some developer who want more efficiency pipelined-write.