@maysamyabandeh 2019-09-26
Did you try the benchmark with --transaction_db which does provide immutable snapshot?