@Little-Wallace 2019-09-27
Thanks your suggestion, I have tried it. 
I agree that the write performance of transaction_db is pretty good.  It is amazing for me.
But there is still other reason which make us not able to use transaction_db this feature, such as it cost more time on `RandomSeek`, and usage of the interface of transaction_db may be a little different from normal DB.  
In the future, we  would try transaction_db in our project,  but we also hope improve write performance with normal DB.