@maysamyabandeh 2019-09-27
Thanks for trying it out. If you look at db_bench_tool.cc, when you pass --transaction_db it opens it with TransactionDB (with certain configs) but then assigns the pointer to DB*, meaning that in the rest of the benchmark only DB:: interfaces are used. This is actually number 1 question in the FAQ of write prepared wiki page: https://github.com/facebook/rocksdb/wiki/unordered_write#faq
If you do see a problem with `RandomSeek` with `TransactionDB` please do bring it up and we shall investigate.