@yiwu-arbug 2019-09-27
@maysamyabandeh Do you know how does `TransactionDB` (after cast into `DB`) compare to vanilla `DBImpl` in general? Do you have users using `TransactionDB` in this way? Thanks.