@yiwu-arbug 2019-09-27
@maysamyabandeh btw, does `TransactionDB` support writing without WAL? My understanding is it heavy rely on WAL, but want to double check with you.