@Little-Wallace 2019-09-29
@maysamyabandeh  when `BenchTool` open one DB with transaction_db, and use `DB::NewIterator` rather than `Transaction::GetIterator` to seek one key, how it could provide consistency read?