@maysamyabandeh 2019-10-17
@yiwu-arbug TransactionDB should be exactly the same as DB when used with the same API. WAL is mandatory when ::Prepare is called which is a TransactionDB specific API. When used with the standard DB API, WAL could be disabled similarly to when using DBImpl.