@maysamyabandeh 2019-10-17
@Little-Wallace One need to define consistent reads to be more clear. When calling TransactionDB::NewIterator it provides exactly the same guarantees as DBImpl::NewIterator. So whatever your definition of consistency is, it will be compromised when switched to TransactionDB.